// src/pages/LocalFood/types/localFood.types.ts
export type localFoodType = {
  local_food_id: number;
  title: string;
  content: string;
  sub_title: string;
  main_thumb_url: string;
  content_url: string;
  lcc_address: string;
  keyword: string;
  create_at: string;
  update_at: string;
  rel_rest_name: string;
  rel_rest_tel: string;
  view_count: number;
  rel_rest_address: string;
};